Feature mes 3949 implement catbe behaviour map (#939)
* add  CATBE behaviour map for vehicle checks with placeholder message

* modify page navigation to use cat be

* fix scss selector name on pass-finalisation page

* remove testing placeholder on office page

* Rename incorrectly named file. Add and refactor tests for behaviour map. Use lodash on selector

* new selector unit tests, and remove console.log

* remove xdescribe from test
