+++ src/pages/debrief/cat-be/__tests__/debrief.cat-be.selector.spec.ts
@@ -7,12 +7,50 @@ import { CatBEUniqueTypes } from '@dvsa/mes-test-schema/categories/BE';
 import {
     displayDrivingFaultComments,
     getManoeuvreFaults,
+    vehicleChecksExist,
   } from '../debrief.cat-be.selector';
 import { CompetencyOutcome } from '../../../../shared/models/competency-outcome';
 
 describe('debriefSelector', () => {
   const manoeuvres: CatBEUniqueTypes.Manoeuvres = {};
 
+  describe('vehicleChecksExist', () => {
+    it('should return false if there are no vehicle checks entered', ()  => {
+      const emptyVehicleChecks: CatBEUniqueTypes.VehicleChecks = { showMeQuestions: [{}], tellMeQuestions: [{}] };
+      const result = vehicleChecksExist(emptyVehicleChecks);
+      expect(result).toBeFalsy();
+    });
+
+    it('should return true if there are vehicle checks (showMeQuestions) entered', ()  => {
+      const populatedVehicleChecks: CatBEUniqueTypes.VehicleChecks =
+        {
+          showMeQuestions: [{ code: 'S01', description: 'test s01', outcome: 'P' }],
+          tellMeQuestions: [{}],
+        };
+      const result = vehicleChecksExist(populatedVehicleChecks);
+      expect(result).toBeTruthy();
+    });
+
+    it('should return true if there are vehicle checks (tellMeQuestions) entered', ()  => {
+      const populatedVehicleChecks: CatBEUniqueTypes.VehicleChecks =
+        {
+          showMeQuestions: [{}],
+          tellMeQuestions: [{ code: 'T01', description: 'test T01', outcome: 'DF' }],
+        };
+      const result = vehicleChecksExist(populatedVehicleChecks);
+      expect(result).toBeTruthy();
+    });
+    it('should return false if there are vehicle checks but no outcome selected', ()  => {
+      const populatedVehicleChecks: CatBEUniqueTypes.VehicleChecks =
+        {
+          showMeQuestions: [{ code: 'S01', description: 'test s01', outcome: null }],
+          tellMeQuestions: [{ code: 'T01', description: 'test T01', outcome: null }],
+        };
+      const result = vehicleChecksExist(populatedVehicleChecks);
+      expect(result).toBeFalsy();
+    });
+
+  });
   describe('displayDrivingFaultComments', () => {
     let localDangerousFaults: DangerousFaults = {};
     let localSeriousFaults: SeriousFaults = {};
