+++ src/pages/office/cat-be/__tests__/office.cat-be.page.spec.ts
@@ -65,7 +65,7 @@ import { NavigationStateProviderMock } from '../../../../providers/navigation-st
 import { SetActivityCode } from '../../../../modules/tests/activity-code/activity-code.actions';
 import { TestCategory } from '../../../../shared/models/test-category';
 
-xdescribe('OfficePage', () => {
+describe('OfficePage', () => {
   let fixture: ComponentFixture<OfficeCatBEPage>;
   let component: OfficeCatBEPage;
   let navController: NavController;
