+++ src/pages/office/cat-be/office.cat-be.page.ts
@@ -78,6 +78,7 @@ import {
   getUncoupleRecoupleFaultAndComment,
   getVehicleCheckSeriousFaults,
   displayDrivingFaultComments,
+  vehicleChecksExist,
   getDrivingFaults,
   getEyesightTestSeriousFaultAndComment,
   getDangerousFaults,
@@ -105,7 +106,7 @@ import {
   CommentSource,
 } from '../../../shared/models/fault-marking.model';
 import { OutcomeBehaviourMapProvider } from '../../../providers/outcome-behaviour-map/outcome-behaviour-map';
-import { behaviourMap } from '../office-behaviour-map';
+import { behaviourMap } from '../office-behaviour-map.cat-be';
 import { ActivityCodeModel, activityCodeModelList } from '../components/activity-code/activity-code.constants';
 import { CompetencyOutcome } from '../../../shared/models/competency-outcome';
 import { startsWith } from 'lodash';
@@ -141,6 +142,7 @@ interface OfficePageState {
   displayDrivingFault$: Observable<boolean>;
   displaySeriousFault$: Observable<boolean>;
   displayDangerousFault$: Observable<boolean>;
+  displayVehicleChecks$: Observable<boolean>;
   identification$: Observable<Identification>;
   independentDriving$: Observable<IndependentDriving>;
   candidateDescription$: Observable<string>;
@@ -331,6 +333,15 @@ export class OfficeCatBEPage extends BasePageComponent {
           select(getDangerousFaults))),
         map(([outcome, dangerousFault]) =>
           this.outcomeBehaviourProvider.isVisible(outcome, 'faultComment', dangerousFault)),
+      ),
+      displayVehicleChecks$: currentTest$.pipe(
+          select(getTestOutcome),
+          withLatestFrom(currentTest$.pipe(
+            select(getTestData))),
+            map(([outcome, data]) =>
+            this.outcomeBehaviourProvider.isVisible(outcome,
+              'vehicleChecks',
+              vehicleChecksExist(data.vehicleChecks))),
       ),
       candidateDescription$: currentTest$.pipe(
         select(getTestSummary),
